pyproject.toml
[tool.poetry]
name = "chatgpt-python"
version = "0.1.0"
description = ""
authors = ["arturj9 <arturjardel9@gmail.com>"]
readme = "README.md"
packages = [{include = "chatgpt_python"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.26.5"
python-dotenv = "^0.21.1"
gtts = "^2.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
